pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "google-generativeai>=0.8.6",
    "gtts>=2.5.4",
    "langdetect>=1.0.9",
    "pydantic>=2.12.5",
    "uvicorn[standard]>=0.44.0",
]
